Conflicts:
	src/main/java/com/universitaria/atelier/web/jpa/Estado.java
	src/main/java/com/universitaria/atelier/web/jpa/Material.java
	src/main/java/com/universitaria/atelier/web/jpa/Ordencompradeta.java
	src/main/java/com/universitaria/atelier/web/jpa/Prenda.java
	src/main/java/com/universitaria/atelier/web/jpa/Produccion.java
	src/main/java/com/universitaria/atelier/web/jpa/Producciondeta.java
	src/main/java/com/universitaria/atelier/web/jpa/Stockmateriales.java
	src/main/java/com/universitaria/atelier/web/jpa/Talla.java
	src/main/resources/META-INF/persistence.xml

+++ src/main/java/com/universitaria/ateliermaven/ejb/compras/DetalleRequerimientoEJB.java
@@ -73,7 +73,7 @@ public class DetalleRequerimientoEJB extends AbstractFacade<Requestdeta> {
                 rqDeta.setMaterialId(em.find(Material.class, Integer.parseInt(mat.getMaterialId())));
                 rqDeta.setRequestDetaFecha(new Timestamp(System.currentTimeMillis()));
                 rqDeta.setUsuarioId(user.getUsuarioId());
-                rqDeta.setRequestDetaCantidad(Float.parseFloat(mat.getCantidad()));
+                rqDeta.setRequestDetaCantidad(Double.parseDouble(mat.getCantidad()));
                 em.persist(rqDeta);                
             }   
             return  true;
@@ -92,7 +92,7 @@ public class DetalleRequerimientoEJB extends AbstractFacade<Requestdeta> {
                 for (MaterialRequerimientoUtil mat : listMaterial) {                
                     if (rqDeta.getMaterialId().getMaterialId()==Integer.parseInt(mat.getMaterialId())) {
                         flag++;
-                        rqDeta.setRequestDetaCantidad(Float.parseFloat(mat.getCantidad()));
+                        rqDeta.setRequestDetaCantidad(Double.parseDouble(mat.getCantidad()));
                         listMaterial.remove(mat);
                         break;
                     }
